+++ src/backend/impress/settings.py
@@ -239,6 +239,8 @@ class Base(Configuration):
             ("fr-fr", "Français"),
             ("de-de", "Deutsch"),
             ("nl-nl", "Nederlands"),
+            ("es-es", "Spanish"),
+            ("cn-cn", "Chinese"),
         )
     )